requirements.txt
pypdf2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.0
pandas>=1.5.0
python-magic>=0.4.27
chardet>=5.0.0
python-pptx>=0.6.21
